requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
playwright==1.46.0
beautifulsoup4==4.12.3
lxml==5.3.0
psycopg2-binary